# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Receitas Culinárias API",
        "description": "API de cadastro e gestão de receitas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API local"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cadastro de usuário",
                "description": "Cria um novo usuário e retorna o token de acesso.",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "login",
                                    "senha",
                                    "senha_confirmation"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Maria Silva"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "maria"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    },
                                    "senha_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Autentica e retorna o token Bearer.",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "senha"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "maria"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado"
                    },
                    "400": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logoff",
                "description": "Invalida o token do usuário autenticado.",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Logout realizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Usuário autenticado",
                "description": "Retorna os dados do usuário logado.",
                "operationId": "99e9c2c91427f6bf63c2f0ef765a768f",
                "responses": {
                    "200": {
                        "description": "Dados do usuário"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/categorias": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar categorias",
                "description": "Retorna todas as categorias para uso em formulários de receita.",
                "operationId": "66f747250d1c2a1a99b78b5ffc76ad86",
                "responses": {
                    "200": {
                        "description": "Lista de categorias"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/receitas": {
            "get": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Listar e pesquisar receitas",
                "description": "Lista receitas do usuário com paginação. Filtros: q (busca em nome/ingredientes/modo_preparo), categoria_id.",
                "operationId": "c88a73109effc476cc2051c29225db00",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Termo de busca",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categoria_id",
                        "in": "query",
                        "description": "ID da categoria",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de receitas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Cadastrar receita",
                "description": "Cria uma nova receita para o usuário autenticado.",
                "operationId": "bd75738047a6505675761c40d7872098",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_categorias",
                                    "nome",
                                    "tempo_preparo_minutos",
                                    "porcoes",
                                    "modo_preparo",
                                    "ingredientes"
                                ],
                                "properties": {
                                    "id_categorias": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nome": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Bolo de chocolate"
                                    },
                                    "tempo_preparo_minutos": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "porcoes": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "modo_preparo": {
                                        "type": "string",
                                        "example": "Misture os ingredientes..."
                                    },
                                    "ingredientes": {
                                        "type": "string",
                                        "example": "Farinha, ovos, chocolate..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Receita criada"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/receitas/{id}": {
            "get": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Exibir receita",
                "description": "Retorna uma receita do usuário. Usado para visualização e impressão.",
                "operationId": "ace250c0ff897d66c0fdb45375c5ba32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receita"
                    },
                    "404": {
                        "description": "Receita não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Editar receita",
                "description": "Atualiza uma receita do usuário. Todos os campos são opcionais.",
                "operationId": "238ffb6d440a42029d2f4d485cad0040",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_categorias": {
                                        "type": "integer"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "maxLength": 45
                                    },
                                    "tempo_preparo_minutos": {
                                        "type": "integer"
                                    },
                                    "porcoes": {
                                        "type": "integer"
                                    },
                                    "modo_preparo": {
                                        "type": "string"
                                    },
                                    "ingredientes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receita atualizada"
                    },
                    "400": {
                        "description": "Erro de validação"
                    },
                    "404": {
                        "description": "Receita não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Excluir receita",
                "description": "Exclui uma receita do usuário.",
                "operationId": "2ef390e37df948d92ff387726318235a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receita excluída"
                    },
                    "404": {
                        "description": "Receita não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token obtido em POST /login ou POST /register. Use: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Cadastro, login e logoff"
        },
        {
            "name": "Categorias",
            "description": "Listagem de categorias"
        },
        {
            "name": "Receitas",
            "description": "CRUD de receitas do usuário"
        }
    ]
}